+++ lib/assent/jwt_adapter/jose.ex
@@ -2,6 +2,10 @@ defmodule Assent.JWTAdapter.JOSE do
   @moduledoc """
   JWT adapter module for parsing JSON Web Tokens with JOSE.
 
+  You can append options to the configuration:
+
+      jwt_adapter: {Assent.JWTAdapter.JOSE, [...]}
+
   See `Assent.JWTAdapter` for more.
   """
   alias Assent.JWTAdapter