+++ test/test_helper.exs
@@ -1,3 +1,5 @@
+Application.put_env(:assent, :http_adapter, Assent.HTTPAdapter.Httpc)
+
 Logger.configure(level: :warning)
 
 ExUnit.start()