+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v0.2.8 (TBA)
 
+`Req` will be used by default if available in your project, otherwise `:httpc` will be used.
+
+- `Req` HTTP adapter added
+- `Req` supported by default as HTTP client
+- Global application config support for HTTP and JWT adapters
 - More expressive errors now including the whole HTTP response where applicable
 
 ## v0.2.7 (2023-09-12)